requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.13.3
blinker==1.8.2
cachetools==5.5.1
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.54.1
gast==0.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.49.1
google-auth-httplib2==0.2.0
google-genai==1.67.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numba==0.60.0
numpy==2.0.2
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.41.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tenacity==9.1.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
traitlets==5.14.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.6.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.0.4
wrapt==1.17.2
wsproto==1.3.2
xgboost==2.1.1
